requirements.txt
praw
transformers
torch
requests 
google-generativeai